FEATURE: Restore full authentication system (Track B Complete)
✨ Track B Complete: Authentication System Integration (1 hour)

Changes:
- Restored complete authentication App.tsx from backup
- Fixed isLoading property reference in AuthRouter
- Backed up professional landing page to App.landing.tsx
- Full JWT authentication system now active

Features Restored:
- User registration and login workflows
- JWT token management with AsyncStorage
- Protected route navigation
- Authentication context and state management
- Login/Register screens with validation
- Logout functionality and token refresh

Next: Track C - Core Lecture Creation (MVP functionality)
